Chunk 10: split FinanceLedger.jsx (M2)
Extraction-only, no behaviour change:
- financeLedgerUtils.js: VIEWS/VIEW_LABELS/YEARS + fmtDate/fmtAmount/isEligible
- FinanceLedgerControls.jsx: view selector, account/category/group/event
  picker, year selector
- FinanceLedgerTable.jsx: sortable transactions table with running balance,
  brought-forward rows, and totals footer

FinanceLedger.jsx 892 -> 395 lines.
